move rundisclaimed subcommand logic to separate package
