+++ cmd/launcher/main.go
@@ -15,8 +15,8 @@ import (
 	"github.com/kolide/kit/env"
 	"github.com/kolide/kit/logutil"
 	"github.com/kolide/kit/version"
-	"github.com/kolide/launcher/ee/allowedcmd"
 	"github.com/kolide/launcher/ee/control/consumers/remoterestartconsumer"
+	"github.com/kolide/launcher/ee/disclaim"
 	"github.com/kolide/launcher/ee/tuf"
 	"github.com/kolide/launcher/ee/watchdog"
 	"github.com/kolide/launcher/pkg/contexts/ctxlog"
@@ -220,7 +220,7 @@ func runSubcommands(systemMultiSlogger *multislogger.MultiSlogger) error {
 	case "query-windowsupdates":
 		run = runQueryWindowsUpdates
 	case "rundisclaimed":
-		run = allowedcmd.RunDisclaimed
+		run = disclaim.RunDisclaimed
 	default:
 		return fmt.Errorf("unknown subcommand %s", os.Args[1])
 	}
