+++ ee/disclaim/run_disclaimed_other.go
@@ -1,7 +1,7 @@
 //go:build !darwin
 // +build !darwin
 
-package allowedcmd
+package disclaim
 
 import (
 	"errors"
